# place2connect-api/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  handlers.SignInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.SignUpInput:
    properties:
      email:
        type: string
      firstName:
        minLength: 3
        type: string
      friends:
        items:
          $ref: '#/definitions/models.User'
        type: array
      impressions:
        type: integer
      isAdmin:
        type: boolean
      isVerified:
        type: boolean
      lastName:
        minLength: 3
        type: string
      location:
        minLength: 3
        type: string
      occupation:
        minLength: 3
        type: string
      password:
        minLength: 8
        type: string
      photo:
        type: string
      picturePath:
        minLength: 3
        type: string
      provider:
        description: PasswordConfirm string        `json:"passwordConfirm" validate:"required,min=8"`
        type: string
      viewedProfile:
        type: integer
    required:
    - firstName
    - lastName
    - location
    - occupation
    - password
    - picturePath
    type: object
...
  handlers.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      firstName:
        type: string
      friends:
        items:
          $ref: '#/definitions/handlers.UserResponse'
        type: array
      id:
        type: integer
      impressions:
        type: integer
      isAdmin:
        type: boolean
      isVerified:
        type: boolean
      lastName:
        type: string
      location:
        type: string
      occupation:
        type: string
      photo:
        type: string
      picturePath:
        type: string
      provider:
        type: string
      role:
        type: string
      updated_at:
        type: string
      viewedProfile:
        type: integer
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      firstName:
        type: string
      friends:
        items:
          $ref: '#/definitions/models.User'
        type: array
      id:
        type: integer
      impressions:
        type: integer
      isAdmin:
        type: boolean
      lastName:
        type: string
      location:
        type: string
      occupation:
        type: string
      password:
        type: string
      photo:
        type: string
      picturePath:
        type: string
      provider:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      verified:
        type: boolean
      viewedProfile:
        type: integer
    type: object
info:
  contact:
    email: contact@owonafx.com
    name: PLACE2CONNECT API Support
  description: PLACE2CONNECT Docs.
  license:
    name: pace2connect 2.0
  termsOfService: http://swagger.io/terms/
  title: PLACE2CONNECT API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Signin/Login  a user to the API.
      parameters:
      - description: Sign In Input
        in: body
        name: signInput
        required: true
        schema:
          $ref: '#/definitions/handlers.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: answer
          schema:
            type: string
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      summary: login a user to the API.
      tags:
      - User
  /api/v1/auth/logout:
    get:
      consumes:
      - application/json
      description: Logout  of the the API.
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Logout  of the the API.
      tags:
      - User
  /api/v1/auth/refresh:
    get:
      consumes:
      - application/json
      description: refresh the current TOKEN.
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "422":
          description: Status UnprocessableEntity
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: refresh the current TOKEN
      tags:
      - User
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Signup/Register a new user.
      parameters:
      - description: SignInput
        in: body
        name: signInput
        required: true
        schema:
          $ref: '#/definitions/handlers.SignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Status BadRequest
          schema:
            type: string
        "403":
          description: Status Forbidden
          schema:
            type: string
        "409":
          description: Status Conflict
          schema:
            type: string
        "502":
          description: Status BadGateway
          schema:
            type: string
      summary: create a new user
      tags:
      - User
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"